Start moving filesystem sandbox to own folder
