+++ @blaxel/core/src/sandbox/index.ts
@@ -6,7 +6,7 @@ export {
   getProcess, getProcessByIdentifier, getProcessByIdentifierLogs, getProcessByIdentifierLogsStream, PortMonitorRequest, postNetworkProcessByPidMonitor, postProcess, ProcessKillRequest,
   ProcessRequest, ProcessResponse, putFilesystemByPath, SuccessResponse
 } from "./client/index.js";
+export * from "./filesystem/index.js";
 export * from "./sandbox.js";
-
 // Re-export everything from client except ClientOptions to avoid conflict
